refactor(Group) : 그룹 관련 VO를 vo 패키지로 이동하여 가시성 개선
- `GroupName`과 같은 엔티티 관련 VO를 `group.domain.vo` 패키지로 분리
- 엔티티와 VO의 역할을 명확히 구분하여 코드 가독성 및 유지보수성 향상
